Fast API Server 1.0 ===================
Description:
Created a server for posts and users and implemented basic crud operations and
created a database using postgress and uploaded data to it locally. Added for creating
tables if table does not exist. Created basic hashing logic using bcrypt and passlib
and hashed the passwords in the database

Testing:
Testing done completely using postman for api testing

Version - 1.0